Switching testing database to the dbtest package
The latest version of mgo includes a dbtest package. This package
manages an instance of MongoDB and provides tools to wipe the DB
before each test run. Hopefully, this will clear up the testing issues
we saw before.

I have adjusted the travis yml becase the new test suite requires that
MongoDB *NOT* be running.