Changes to how the fact creation happens for observations
This now uses EffectiveDateTime instead of EffectivePeriod. This is
better for a few reasons:
* The fact creation would panic if an Effective Period was not
present. Now it should work regardless of whether EffectiveDateTime is
there or not.
* For most Observations EffectiveDateTime makes more sense than using
a period.

There were no corresponding tests to update here and I'm not really
concerned since we will be tossing this code when we refactor the
pipeline.