Support for rollover scheduling
Will look back the configured rollover delay days to see if there was a huddle.  If so, it will look for unreviewed patients and add them to the first huddle we schedule.

This also fixes a potential bug where today's huddle might be modified during or after it's in progress.